Create a plugin file.
WordPress looks for a file in the root of the plugin directory that contains specific file headers. These headers define the plugin name, URI, description, version, author, etc.

Adding this file with the appropriate headers is the minimum requirement to get a plugin to show in the WordPress plugin listing screen. You can now go into WordPress and activate the plugin.

Note that the folder name and main plugin file have the same slug. While it is possible to have a file with a different slug, using the same slug is a common practice. The directory slug and text domain should always be the same.