Register a custom post type.
We prefixed our post type with `rtut-` and defined it in a class constant. Defining the internal post type string in one place makes it really easy to refactor if we ever rename it in the future.

All post type labels are setup for translation and use the appropriate text domain. The recipe post type is public and has the post type archive page enabled. We've assigned a custom dashicon (https://developer.wordpress.org/resource/dashicons/#carrot) and customized the URL for recipes to live at `/recipes`. We've enabled support for titles, the editor, featured images, and revisions.

Our post type is registered on the `init` action hook. It is also registered in our activation hook just before we flush the rewrite rules. We have to flush the rewrite rules on an activation hook so that our pubic custom post type URLs will work automatically. If we did not do this, the post type would exist in the back end, but none of the front-end URLs for the post type would work (unless your permalink structure was using the `?p=1` structure).