Register our taxonomies: Courses and Cuisines
We want the URL structure to be `recipes/courses` and `recipes/cuisines`, so we have to register our taxonomies BEFORE our post type. Otherwise, the rewrite rules would be in the wrong order and the post type rewrite rule would be hit first and our taxonomies would never route.

It is possible to associate our post type when creating our taxonomies or associate our taxonomies when creating our post type. However, I've opted to use the `register_taxonomy_for_object_type` function and handle this independently.